move legacy agenda to the drools-runtime module
