+++ drools-traits/src/test/java/org/drools/traits/UseOfRuleFlowGroupPlusLockOnTest.java
@@ -1,6 +1,6 @@
 package org.drools.traits;
 
-import org.drools.core.common.DefaultAgenda;
+import org.drools.runtime.agenda.DefaultAgenda;
 import org.drools.core.reteoo.ReteDumper;
 import org.junit.Assert;
 import org.junit.Test;
